spdx-converter/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
isodate==0.6.0
ply==3.11
pyparsing==2.4.7
python-dateutil==2.8.1
PyYAML==5.3.1
rdflib==5.0.0
requests==2.23.0
six==1.14.0
spdx-tools==0.6.1
urllib3==1.25.9
xmltodict==0.12.0
